partial refactor ( doesn't compile )
